// src/types/studentObservations.ts
export interface StudentObservations {
  id: number;
  student_id: number;
  obs_reading_writing: number;
  obs_mirror_numbers: number;
  obs_left_right_confusion: number;
  obs_sequence_difficulty: number;
  obs_disorganized_work: number;
  obs_inattention_detail: number;
  obs_sustained_attention: number;
  obs_doesnt_listen: number;
  obs_task_organization: number;
  obs_loses_belongings: number;
  obs_distracted_stimuli: number;
  obs_forgetful: number;
  obs_excess_hand_foot: number;
  obs_gets_up_from_seat: number;
  obs_running_jumping: number;
  obs_talks_excessively: number;
  obs_difficulty_quiet: number;
  obs_driven_by_motor: number;
  obs_impulsive_answers: number;
  obs_difficulty_waiting: number;
  obs_interrupts_others: number;
  obs_easily_angered: number;
  obs_argues: number;
  obs_defies_adults: number;
  obs_annoys_others: number;
  obs_aggressive: number;
  obs_spiteful: number;
  obs_blames_others: number;
  obs_breaks_property: number;
  obs_incomplete_homework: number;
  obs_frequent_absences: number;
  obs_neglected_appearance: number;
  obs_uses_profanity: number;
  obs_takes_belongings: number;
  obs_forgets_materials: number;
  obs_appears_sad: number;
}

export interface StudentObservationsInput {
  obs_reading_writing: boolean;
  obs_mirror_numbers: boolean;
  obs_left_right_confusion: boolean;
  obs_sequence_difficulty: boolean;
  obs_disorganized_work: boolean;
  obs_inattention_detail: boolean;
  obs_sustained_attention: boolean;
  obs_doesnt_listen: boolean;
  obs_task_organization: boolean;
  obs_loses_belongings: boolean;
  obs_distracted_stimuli: boolean;
  obs_forgetful: boolean;
  obs_excess_hand_foot: boolean;
  obs_gets_up_from_seat: boolean;
  obs_running_jumping: boolean;
  obs_talks_excessively: boolean;
  obs_difficulty_quiet: boolean;
  obs_driven_by_motor: boolean;
  obs_impulsive_answers: boolean;
  obs_difficulty_waiting: boolean;
  obs_interrupts_others: boolean;
  obs_easily_angered: boolean;
  obs_argues: boolean;
  obs_defies_adults: boolean;
  obs_annoys_others: boolean;
  obs_aggressive: boolean;
  obs_spiteful: boolean;
  obs_blames_others: boolean;
  obs_breaks_property: boolean;
  obs_incomplete_homework: boolean;
  obs_frequent_absences: boolean;
  obs_neglected_appearance: boolean;
  obs_uses_profanity: boolean;
  obs_takes_belongings: boolean;
  obs_forgets_materials: boolean;
  obs_appears_sad: boolean;
}